pyproject.toml
[project]
name = "auth"
version = "0.0.1"
authors = [{ name = "Ramiro e Vinicius", email = "fulano@pucrs.com" }]
description = "A FastAPI project for user management"
readme = "README.md"
requires-python = ">=3.12"
keywords = ["auth"]
dependencies = [
    "fastapi~=0.111",
    "requests~=2.26",
    "python-multipart",
    "loguru~=0.7",
    "pydantic[email]>=2,<3",
    "pydantic-settings",
    "uvicorn==0.30.3",
    "fastapi-cli==0.0.4",
    "httpx==0.27.0",
    "pymongo==4.8.0",
    "python-dotenv==1.0.1",
    "rich==13.7.1",
    "starlette==0.37.2",
    "toml==0.10.2",
    "typer==0.12.3",
    "typing-extensions==4.12.2",
    "uvloop==0.19.0",
    "websockets==12.0",
    "google-auth==2.34.0",
    "PyJWT==2.9.0",
    "cryptography",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
addopts = [
    "--import-mode=importlib",
    "--cov=src",
    "--cov-report=html",
    "--disable-warnings",
]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
